refactor(Proof Price Tag Assistant): replace ContributionAssistantLabelList with ContributionAssistantLabelCard component
